+++ CHANGELOG.md
@@ -6,12 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Updated
 - Improve performance across the board by:
   - removing `Hash#fetch`
   - using `Array#hash`
   - avoiding multi-layer hash lookups for multi-argument methods
   - optimizing for truthy results
 - Add `dry-core` to benchmarks in README
+### Fixed
+- Fix usage on module singleton classes
+- Fix usage on module which would be extended by other classes
 
 ## [1.0.0] - 2021-06-24
 ### Added